requirements.txt
cmake==4.1.0
tensordict
accelerate
codetiming
datasets==4.0.0
hydra-core==1.3.2
peft==0.17.1
pandas==2.3.2
pylatexenc==2.10
torchdata==0.11.0
wandb
uvloop==0.21.0
tensorboard==2.20.0
mathruler==0.1.0
numba==0.61.2
ray[default]
cbor2==5.7.0
pybase64
openai-harmony==0.0.4
modelscope==1.29.2
transformers==4.55.0
uvicorn==0.35.0
openai==1.105.0
py-cpuinfo==9.0.0
gguf
aiofiles==24.1.0
aiohttp==3.12.15
anyio==4.10.0
attrs==25.3.0
blake3==1.0.5
cachetools==6.2.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.10.2
Cython==3.1.2
dill==0.3.8
dnspython==2.7.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
grpcio==1.74.0
hf-xet==1.1.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistral_common==1.8.4
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
ninja==1.13.0
omegaconf==2.3.0
orjson==3.11.3
outlines_core==0.2.10
partial-json-parser==0.2.1.1.post6
pillow==11.3.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.2
pyzmq==27.0.2
Quart==0.20.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.4
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.27.1
safetensors==0.6.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.35.2
setproctitle==1.3.6
setuptools-scm==9.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.47.3
sympy==1.14.0
tensorboard-data-server==0.7.2
tiktoken==0.11.0
tokenizers==0.21.4
torchvision==0.22.1
tqdm==4.67.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xgrammar==0.1.21
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
